test(opencode): isolate shared test helpers
