+++ tests/adapters/opencode/test_setup.py
@@ -17,7 +17,7 @@ from band.runtime.tools import CONTACT_TOOL_NAMES, MEMORY_TOOL_NAMES
 from band.testing import FakeAgentTools
 
 
-from .conftest import (
+from .helpers import (
     FakeMCPBackend,
     FakeOpencodeClient,
     _make_fake_mcp_backend_factory,
