+++ tests/adapters/opencode/test_turns.py
@@ -17,7 +17,7 @@ from band.testing import FakeAgentTools
 from tests.adapters.usage_events import recorded_usage_payloads
 
 
-from .conftest import (
+from .helpers import (
     AnyHTTPStatusError,
     FakeOpencodeClient,
     MentionEnforcingTools,
